make the network handler for fabric generic. this allows the developer to plugin different ways to manage the fabric namespaces
