+++ integration/nwo/token/factory.go
@@ -11,6 +11,7 @@ import (
 	fabric2 "github.com/hyperledger-labs/fabric-smart-client/integration/nwo/fabric"
 	"github.com/hyperledger-labs/fabric-smart-client/integration/nwo/orion"
 	"github.com/hyperledger-labs/fabric-token-sdk/integration/nwo/token/fabric"
+	"github.com/hyperledger-labs/fabric-token-sdk/integration/nwo/token/fabric/cc"
 	orion2 "github.com/hyperledger-labs/fabric-token-sdk/integration/nwo/token/orion"
 )
 
@@ -26,7 +27,7 @@ func (p *platformFactory) Name() string {
 
 func (p *platformFactory) New(ctx api.Context, t api.Topology, builder api.Builder) api.Platform {
 	tp := NewPlatform(ctx, t, builder)
-	tp.AddNetworkHandler(fabric2.TopologyName, fabric.NewNetworkHandler(tp, builder))
+	tp.AddNetworkHandler(fabric2.TopologyName, fabric.NewNetworkHandler(tp, builder, cc.NewDefaultGenericBackend(tp)))
 	tp.AddNetworkHandler(orion.TopologyName, orion2.NewNetworkHandler(tp, builder))
 	return tp
 }
